% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{ca_models}
\alias{ca_models}
\title{List available climate models}
\usage{
ca_models(activity = NULL)
}
\arguments{
\item{activity}{character; "WRF" or "LOCA2" or NULL for both.}
}
\value{
character vector of model names (source_id values)
}
\description{
List available climate models
}
