% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{cae_scenarios}
\alias{cae_scenarios}
\title{List available scenarios (SSPs)}
\usage{
cae_scenarios(activity = NULL, model = NULL)
}
\arguments{
\item{activity}{character; "WRF" or "LOCA2" or NULL for both.}

\item{model}{character; filter to a specific GCM. NULL for all.}
}
\value{
character vector of experiment_id values
}
\description{
List available scenarios (SSPs)
}
