% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{cae_variables}
\alias{cae_variables}
\title{List available variables}
\usage{
cae_variables(activity = NULL, timescale = NULL)
}
\arguments{
\item{activity}{character; "WRF" or "LOCA2" or NULL for both.}

\item{timescale}{character; "1hr", "day", "mon", or NULL for all.}
}
\value{
character vector of variable_id values
}
\description{
List available variables
}
