% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caladaptaer-package.R
\docType{package}
\name{caladaptaer-package}
\alias{caladaptaer}
\alias{caladaptaer-package}
\title{caladaptaer: Access Cal-Adapt Climate Data from R}
\description{
Provides native R access to the Cal-Adapt Analytics Engine, California's repository of downscaled CMIP6 climate projections. Supports both WRF dynamically downscaled (hourly, 3/9/45 km) and LOCA2-Hybrid statistically downscaled (daily, 3 km) datasets. Reads cloud-optimized Zarr stores directly from AWS S3 via GDAL, with no Python dependency.

Native R access to the Cal-Adapt Analytics Engine, California's
repository of downscaled CMIP6 climate projections. Reads WRF
dynamically downscaled and LOCA2-Hybrid statistically downscaled
data directly from AWS S3 Zarr stores via GDAL.
}
\section{Catalog functions}{

\itemize{
\item \code{\link{cae_catalog}} -- load the data catalog
\item \code{\link{cae_models}} -- list available GCMs
\item \code{\link{cae_scenarios}} -- list available SSPs
\item \code{\link{cae_variables}} -- list available variables
\item \code{\link{cae_search}} -- search the catalog
}
}

\section{Data access functions}{

\itemize{
\item \code{\link{cae_fetch}} -- fetch gridded data as stars object
\item \code{\link{cae_fetch_point}} -- extract time series at a point
}
}

\section{Export functions}{

\itemize{
\item \code{\link{cae_write_cf_netcdf}} -- write CF-standard NetCDF
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lebauerapproach/caladaptaer}
  \item Report bugs at \url{https://github.com/lebauerapproach/caladaptaer/issues}
}

}
\author{
\strong{Maintainer}: Akash B V \email{abv1@bu.edu}

Authors:
\itemize{
  \item David LeBauer \email{dlebauer@gmail.com} (\href{https://orcid.org/0000-0001-7228-053X}{ORCID})
}

}
\keyword{internal}
