% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\docType{data}
\name{.catalog_csv_url}
\alias{.catalog_csv_url}
\title{Cal-Adapt Analytics Engine Data Catalog}
\format{
An object of class \code{character} of length 1.
}
\usage{
.catalog_csv_url
}
\description{
Functions for loading, caching, and querying the Cal-Adapt
intake-ESM catalog. The catalog indexes all Zarr stores on the
CADCAT S3 bucket with CMIP6-style metadata fields.
}
\keyword{datasets}
