% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{.compute_wind_speed}
\alias{.compute_wind_speed}
\title{Compute wind speed from components}
\usage{
.compute_wind_speed(u10, v10)
}
\arguments{
\item{u10}{numeric vector; eastward wind in m/s}

\item{v10}{numeric vector; northward wind in m/s}
}
\value{
numeric vector; wind speed in m/s
}
\description{
Compute wind speed from components
}
\keyword{internal}
