% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.detect_activity}
\alias{.detect_activity}
\title{Detect activity_id from model name and params}
\usage{
.detect_activity(model, timescale, resolution, variable = NULL)
}
\description{
Detect activity_id from model name and params
}
\keyword{internal}
