% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\docType{data}
\name{.wrf_cf_map}
\alias{.wrf_cf_map}
\title{Unit Conversion and CF Standard Formatting}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
.wrf_cf_map
}
\description{
Functions to convert Cal-Adapt WRF variables to CF-standard
NetCDF format. Handles variable renaming, unit conversions, and
metadata stamping.
}
\keyword{datasets}
