% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{.write_cf_nc}
\alias{.write_cf_nc}
\title{Write CF-compliant NetCDF from converted data}
\usage{
.write_cf_nc(converted, time_vals, dims, outfile)
}
\description{
Write CF-compliant NetCDF from converted data
}
\keyword{internal}
