% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.fetch_single}
\alias{.fetch_single}
\title{Read a single variable from S3}
\usage{
.fetch_single(
  variable,
  activity,
  model,
  scenario,
  timescale,
  resolution,
  start_time,
  end_time,
  n_timesteps
)
}
\description{
Read a single variable from S3
}
\keyword{internal}
